Spotlight

+++ TapBand_SRC/Assets/Script/Events/TapEvent.cs
@@ -6,10 +6,10 @@ public delegate void TapEvent(object sender, TapEventArgs e);
 
 public class TapEventArgs : EventArgs
 {
-    public TapEventArgs(float tapStrength)
+    public TapEventArgs(double tapStrength)
     {
         TapStrength = tapStrength;
     }
 
-    public float TapStrength { get; private set; }
+    public double TapStrength { get; private set; }
 }
